+++ frontend/test/pytest/test_buffer_args.py
@@ -78,6 +78,22 @@ class TestReturnValues:
         result2 = order2(data_in)
         assert jnp.allclose(result1[0], result2[1]) and jnp.allclose(result1[1], result2[0])
 
+    @pytest.mark.parametrize("dtype", [(jnp.complex128), (jnp.complex64)])
+    def test_return_complex_scalar(self, dtype):
+        """Complex scalars (and float16)* take a different path when being returned from the
+        compiled function. See `ranked_memref_to_numpy` and `to_numpy` in
+        llvm-project/mlir/python/mlir/runtime/np_to_memref.py.
+
+        * float16 isn't handled by the compiler, so it is not used as a test.
+        """
+
+        @qjit
+        # pylint: disable=missing-function-docstring
+        def return_scalar():
+            return jnp.array(0, dtype=dtype)
+
+        assert jnp.allclose(return_scalar(), complex(0, 0))
+
 
 if __name__ == "__main__":
     pytest.main(["-x", __file__])